added response.isSuccessful before adding to room, faking network request

+++ app/src/main/java/com/example/module_3_lesson_6_hw_2_compose/ui/rooms/RoomListViewModel.kt
@@ -1,9 +1,12 @@
 package com.example.module_3_lesson_6_hw_2_compose.ui.rooms
 
+import android.util.Log
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.example.module_3_lesson_6_hw_2_compose.data.RoomItem
 import com.example.module_3_lesson_6_hw_2_compose.data.RoomsRepository
+import com.example.module_3_lesson_6_hw_2_compose.data.network.API
+import com.example.module_3_lesson_6_hw_2_compose.data.network.RetrofitClient
 import kotlinx.coroutines.flow.SharingStarted
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.map
@@ -22,16 +25,31 @@ class RoomListViewModel(private val roomsRepository: RoomsRepository) : ViewMode
         private const val TIMEOUT_MILLIS = 5_000L
     }
 
+    private val retrofit = RetrofitClient.getClient("https://catfact.ninja/")
+        .create(API::class.java)
+
     suspend fun updateRoomIsLightOn(roomId: Int, isLightOn: Boolean) {
-        val room = roomsRepository.getRoomById(roomId)
-        room?.let {
-            val updatedRoom = it.copy(isLightOn = isLightOn)
-            roomsRepository.updateRoom(updatedRoom)
+        val response = retrofit.fetchSomeData()
+        if (response.isSuccessful) {
+            val room = roomsRepository.getRoomById(roomId)
+            room?.let {
+                val updatedRoom = it.copy(isLightOn = isLightOn)
+                roomsRepository.updateRoom(updatedRoom)
+            }
+            Log.d("MYLOG", response.body()?.fact.toString())
+        } else {
+            Log.d("MYLOG", "Response not successful. Code: ${response.code()}, Message: ${response.message()}")
         }
     }
 
     suspend fun deleteRoomItem(roomItem: RoomItem) {
-        roomsRepository.deleteRoom(roomItem)
+        val response = retrofit.fetchSomeData()
+        if (response.isSuccessful) {
+            roomsRepository.deleteRoom(roomItem)
+            Log.d("MYLOG", response.body()?.fact.toString())
+        } else {
+            Log.d("MYLOG", "Response not successful. Code: ${response.code()}, Message: ${response.message()}")
+        }
     }
 }
